requirements.txt
# Core dependencies
slack_bolt==1.22.0
slack_sdk==3.34.0
python-dotenv==1.0.1

# AWS
boto3==1.37.15
botocore==1.37.15

# LangChain
langchain==0.3.20
langchain-core==0.3.44
langchain-community==0.3.19
langchain-text-splitters==0.3.6

# OpenAI
openai==1.66.3

# Monitoring
langfuse==2.60.0

# HTTP and API dependencies
requests==2.32.3
urllib3==1.26.20

# Required by dependencies
pydantic==2.10.6
pydantic_core==2.27.2
typing_extensions==4.12.2